requirements.txt
docopt==0.6.2
rich==12.6.0
python-dateutil==2.8.2
pyinstaller